+++ src/main/scala/edu/ie3/simona/model/participant/HpModel.scala
@@ -12,7 +12,7 @@ import edu.ie3.simona.model.SystemComponent
 import edu.ie3.simona.model.participant.HpModel.{HpRelevantData, HpState}
 import edu.ie3.simona.model.participant.control.QControl
 import edu.ie3.simona.model.thermal.ThermalGrid.{
-  ThermalEnergyDemand,
+  ThermalDemandWrapper,
   ThermalGridState,
 }
 import edu.ie3.simona.model.thermal.{ThermalGrid, ThermalThreshold}
@@ -132,7 +132,7 @@ final case class HpModel(
   ): (Boolean, Boolean, HpState) = {
 
     // Use lastHpState and relevantData to update state of thermalGrid to the current tick
-    val (demandHouse, demandThermalStorage, currentThermalGridState) =
+    val (thermalDemandWrapper, currentThermalGridState) =
       thermalGrid.energyDemandAndUpdatedState(
         relevantData.currentTick,
         lastHpState.ambientTemperature.getOrElse(
@@ -148,13 +148,12 @@ final case class HpModel(
         lastHpState,
         currentThermalGridState,
         relevantData,
-        demandHouse,
-        demandThermalStorage,
+        thermalDemandWrapper,
       )
 
     // Updating the HpState
     val updatedState =
-      calcState(lastHpState, relevantData, turnOn)
+      calcState(lastHpState, relevantData, turnOn, thermalDemandWrapper)
     (canOperate, canBeOutOfOperation, updatedState)
   }
 
@@ -170,10 +169,8 @@ final case class HpModel(
     *   to current tick updated state of the thermalGrid
     * @param relevantData
     *   Relevant (external) data
-    * @param demandHouse
-    *   ThermalEnergyDemand of the house
-    * @param demandThermalStorage
-    *   ThermalEnergyDemand of the thermal storage
+    * @param thermalDemands
+    *   ThermalEnergyDemand of the house and the thermal storage
     * @return
     *   boolean defining if heat pump runs in next time step, if it can be in
     *   operation and can be out of operation
@@ -182,23 +179,21 @@ final case class HpModel(
       lastState: HpState,
       currentThermalGridState: ThermalGridState,
       relevantData: HpRelevantData,
-      demandHouse: ThermalEnergyDemand,
-      demandThermalStorage: ThermalEnergyDemand,
+      thermalDemands: ThermalDemandWrapper,
   ): (Boolean, Boolean, Boolean) = {
 
-    val (
-      houseHasDemand,
-      heatStorageHasDemand,
-      noThermalStorageOrThermalStorageIsEmpty,
-    ) = determineDemandBooleans(
-      lastState,
-      currentThermalGridState,
-      demandHouse,
-      demandThermalStorage,
+    val demandHouse = thermalDemands.houseDemand
+    val demandThermalStorage = thermalDemands.heatStorageDemand
+
+    val noThermalStorageOrThermalStorageIsEmpty = determineThermalStorageStatus(
+      currentThermalGridState
     )
 
-    val turnHpOn: Boolean =
-      houseHasDemand || heatStorageHasDemand
+    val turnHpOn =
+      (demandHouse.hasRequiredDemand && noThermalStorageOrThermalStorageIsEmpty) ||
+        (demandHouse.hasAdditionalDemand && lastState.isRunning) ||
+        demandThermalStorage.hasRequiredDemand ||
+        (demandThermalStorage.hasAdditionalDemand && lastState.isRunning)
 
     val canOperate =
       demandHouse.hasRequiredDemand || demandHouse.hasAdditionalDemand ||
@@ -206,33 +201,26 @@ final case class HpModel(
     val canBeOutOfOperation =
       !(demandHouse.hasRequiredDemand && noThermalStorageOrThermalStorageIsEmpty)
 
-    (turnHpOn, canOperate, canBeOutOfOperation)
+    (
+      turnHpOn,
+      canOperate,
+      canBeOutOfOperation,
+    )
   }
 
   /** This method will return booleans whether there is a heat demand of house
     * or thermal storage as well as a boolean indicating if there is no thermal
     * storage, or it is empty.
     *
-    * @param lastHpState
-    *   Current state of the heat pump
     * @param updatedGridState
     *   The updated state of the [[ThermalGrid]]
-    * @param demandHouse
-    *   heat demand of the thermal house
-    * @param demandThermalStorage
-    *   heat demand of the thermal storage
     * @return
-    *   First boolean is true, if house has heat demand. Second boolean is true,
-    *   if thermalStorage has heat demand. Third boolean is true, if there is no
-    *   thermalStorage, or it's empty.
+    *   boolean which is true, if there is no thermalStorage, or it's empty.
     */
 
-  private def determineDemandBooleans(
-      lastHpState: HpState,
-      updatedGridState: ThermalGridState,
-      demandHouse: ThermalEnergyDemand,
-      demandThermalStorage: ThermalEnergyDemand,
-  ): (Boolean, Boolean, Boolean) = {
+  private def determineThermalStorageStatus(
+      updatedGridState: ThermalGridState
+  ): Boolean = {
     implicit val tolerance: Energy = KilowattHours(1e-3)
     val noThermalStorageOrThermalStorageIsEmpty: Boolean =
       updatedGridState.storageState.isEmpty || updatedGridState.storageState
@@ -240,11 +228,7 @@ final case class HpModel(
           _.storedEnergy =~ zeroKWh
         )
 
-    val houseDemand =
-      (demandHouse.hasRequiredDemand && noThermalStorageOrThermalStorageIsEmpty) || (lastHpState.isRunning && demandHouse.hasAdditionalDemand)
-    val heatStorageDemand =
-      demandThermalStorage.hasRequiredDemand || (lastHpState.isRunning && demandThermalStorage.hasAdditionalDemand)
-    (houseDemand, heatStorageDemand, noThermalStorageOrThermalStorageIsEmpty)
+    noThermalStorageOrThermalStorageIsEmpty
   }
 
   /** Calculate state depending on whether heat pump is needed or not. Also
@@ -257,6 +241,8 @@ final case class HpModel(
     *   data of heat pump including state of the heat pump
     * @param isRunning
     *   determines whether the heat pump is running or not
+    * @param demandWrapper
+    *   holds the thermal demands of the thermal units (house, storage)
     * @return
     *   next [[HpState]]
     */
@@ -264,6 +250,7 @@ final case class HpModel(
       lastState: HpState,
       relevantData: HpRelevantData,
       isRunning: Boolean,
+      demandWrapper: ThermalDemandWrapper,
   ): HpState = {
     val lastStateStorageQDot = lastState.thermalGridState.storageState
       .map(_.qDot)
@@ -346,13 +333,25 @@ final case class HpModel(
       lastState: HpState,
       setPower: Power,
   ): (HpState, FlexChangeIndicator) = {
-    /* If the setpoint value is above 50 % of the electrical power, turn on the heat pump otherwise turn it off */
+    /* If the set point value is above 50 % of the electrical power, turn on the heat pump otherwise turn it off */
     val turnOn = setPower > (sRated.toActivePower(cosPhiRated) * 0.5)
+
+    val (
+      thermalDemands,
+      _,
+    ) =
+      thermalGrid.energyDemandAndUpdatedState(
+        data.currentTick,
+        lastState.ambientTemperature.getOrElse(data.ambientTemperature),
+        data.ambientTemperature,
+        lastState.thermalGridState,
+      )
 
     val updatedHpState = calcState(
       lastState,
       data,
       turnOn,
+      thermalDemands,
     )
 
     (